+++ src/FederationLib/Enums/Method.php
@@ -7,6 +7,7 @@
     use FederationLib\Methods\Attachments\DeleteAttachment;
     use FederationLib\Methods\Attachments\DownloadAttachment;
     use FederationLib\Methods\Attachments\GetAttachmentInfo;
+    use FederationLib\Methods\Attachments\ListAttachments;
     use FederationLib\Methods\Attachments\UploadAttachment;
     use FederationLib\Methods\Audit\ListAuditLogs;
     use FederationLib\Methods\Audit\ViewAuditEntry;
@@ -96,6 +97,7 @@
         case DOWNLOAD_ATTACHMENT;
         case GET_ATTACHMENT_INFO;
         case DELETE_ATTACHMENT;
+        case LIST_ATTACHMENTS;
 
         /**
          * Handles the request of the method
@@ -133,6 +135,9 @@
                     break;
                 case self::GET_ATTACHMENT_INFO:
                     GetAttachmentInfo::handleRequest();
+                    break;
+                case self::LIST_ATTACHMENTS:
+                    ListAttachments::handleRequest();
                     break;
 
                 case self::LIST_ENTITIES:
@@ -261,6 +266,7 @@
                 preg_match('#^/attachments/([a-fA-F0-9\-]{36})$#', $path) && $requestMethod === 'GET' => Method::DOWNLOAD_ATTACHMENT,
                 preg_match('#^/attachments/([a-fA-F0-9\-]{36})$#', $path) && $requestMethod === 'DELETE' => Method::DELETE_ATTACHMENT,
                 preg_match('#^/attachments/([a-fA-F0-9\-]{36})/info$#', $path) && $requestMethod === 'GET' => Method::GET_ATTACHMENT_INFO,
+                $path === '/attachments' && $requestMethod === 'GET' => Method::LIST_ATTACHMENTS,
                 $path === '/attachments' && ($requestMethod === 'POST' || $requestMethod === 'PUT')  => Method::UPLOAD_ATTACHMENT,
 
                 $path === '/entities' && $requestMethod === 'GET' => Method::LIST_ENTITIES,